refactor: better dir and files structure for the fair algorithm
