+++ otary/image/components/transformer/components/binarizer/utils/fair/constant.py
@@ -1,0 +1,3 @@
+""" """
+
+UNKNOWN_LABEL = 0.5
